Migrate grid, summary list, table and tab to ViewComponents
Convert the remaining data-builder components to ViewComponents,
following the list component pattern: the helper builds the existing
PORO data object (Grid/SummaryList/Table/Tab) and renders a per-brand
component that owns the markup.

- ds_grid / ds_summary_list / ds_table / ds_tab now resolve brand
  components via DesignSystem::Registry.component
- Brand-specific markup (GOV.UK summary-list values, GOV.UK/NHS table
  classes and NHS responsive cells, NHS hidden text) is preserved via
  subclasses that override the relevant private methods
- Data classes under lib/design_system/components stay as the builder
  input; the lib/design_system/*/builders/*.rb PORO builders and their
  requires are removed
- Specs moved under test/*/components
