+++ app/components/design_system/govuk/grid_component.rb
@@ -1,0 +1,6 @@
+module DesignSystem
+  module Govuk
+    class GridComponent < DesignSystem::Generic::GridComponent
+    end
+  end
+end
