+++ app/components/design_system/nhsuk/grid_component.rb
@@ -1,0 +1,6 @@
+module DesignSystem
+  module Nhsuk
+    class GridComponent < DesignSystem::Generic::GridComponent
+    end
+  end
+end
