+++ app/components/design_system/nhsuk/tab_component.rb
@@ -1,0 +1,6 @@
+module DesignSystem
+  module Nhsuk
+    class TabComponent < DesignSystem::Govuk::TabComponent
+    end
+  end
+end
