+++ lib/design_system/generic.rb
@@ -4,7 +4,4 @@
 
 require_relative 'generic/builders/code'
 require_relative 'generic/builders/fixed_elements'
-require_relative 'generic/builders/grid'
 require_relative 'generic/builders/pagination_renderer'
-require_relative 'generic/builders/tab'
-require_relative 'generic/builders/table'
