+++ lib/design_system/govuk.rb
@@ -5,9 +5,6 @@
 require_relative 'govuk/builders/code'
 require_relative 'govuk/builders/fixed_elements'
 require_relative 'govuk/builders/pagination_renderer'
-require_relative 'govuk/builders/summary_list'
-require_relative 'govuk/builders/tab'
-require_relative 'govuk/builders/table'
 
 require_relative 'govuk/form_builder'
