--- lib/design_system/govuk/builders/grid.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-module DesignSystem
-  module Govuk
-    module Builders
-      # GOV UK specific grid implementation
-      # Inherits all functionality from generic builder
-      class Grid < ::DesignSystem::Generic::Builders::Grid
-      end
-    end
-  end
-end
