--- lib/design_system/nhsuk/builders/grid.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-module DesignSystem
-  module Nhsuk
-    module Builders
-      # NHS UK specific grid implementation
-      # Inherits all functionality from generic builder
-      class Grid < ::DesignSystem::Generic::Builders::Grid
-      end
-    end
-  end
-end
