--- lib/design_system/nhsuk/builders/tab.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module DesignSystem
-  module Nhsuk
-    module Builders
-      # This class provides Nhsuk Tab.
-      class Tab < ::DesignSystem::Govuk::Builders::Tab
-      end
-    end
-  end
-end
