+++ test/govuk/components/table_test.rb
@@ -2,7 +2,7 @@ require 'test_helper'
 
 module DesignSystem
   module Govuk
-    module Builders
+    module Components
       # This tests the govuk headings builder
       class TableTest < ActionView::TestCase
         include DesignSystemHelper
